refactor : "프로젝트 정보" 입력 폼과 출력 폼으로 분리
